==> src/sprint_4/03_define_cohorts.R <==
####
# Define cohorts of care come residents (January 2019 and January 2020)
####

library(tidyverse)
library(tidylog)
library(janitor)
library(lubridate)
library(tableone)
library(RColorBrewer)


source("file_paths.R")


# Import data -------------------------------------------------------------

# Import MPI with care home characteristics 
mmpi_monthly <- readRDS(str_c(Rds_data_path, 'sprint_4/mmpisubset_chres_monthly_chchars.Rds')) %>% 
  filter(month >= ymd("2019-01-01"))

# IMD lookup file (based on LSOA11)
imd_19_lookup <- read_csv("../reference_data/IMD_2019_LSOA.csv") %>% 
  select(starts_with('LSOA code'),
         contains('Index of Multiple Deprivation (IMD) Rank'),
         contains('Index of Multiple Deprivation (IMD) Decile'))
names(imd_19_lookup) <- c("LSOA11", "IMD19_RANK", "IMD19_DECILE")

# Long-term condition flag, generated by RW based on diagnoses from the previous 3 years of SUS episodes
# this was done with two index dates: 1 January 2019 and 1 January 2020
ltcs <- read_csv(str_c(raw_data_path, "care home paper/cohorts/cohorts_ltcs_charlson.csv"),
                 col_types = cols(pid = col_character())) %>% 
  select(-uid, -findexdate)

# LSOA to region lookup
lsoa11_to_region <- read_csv("../reference_data/lsoa11_reg20_lu.csv")

# Define cohorts ----------------------------------------------------------

mmpi_monthly <- mmpi_monthly %>% 
  mutate(cohort_2019 = ifelse(month == ymd("2019-01-01"), 1, 0),
         cohort_2020 = ifelse(month == ymd("2020-01-01"), 1, 0),
         age = interval(dob, extractstart) %/% years(1),
         overand65 = ifelse(age >= 65, 1, 0),
         under65_learndis = ifelse(age < 65 & user_learning == 1, 1, 0)) 

# Study params
startdate_2019 <- ymd("2019-01-01")
enddate_2019 <- ymd("2019-06-30")

startdate_2020 <- ymd("2020-01-01")
enddate_2020 <- ymd("2020-06-30")

# Filter for cohort IDs and only keep relevant months
chres_cohort_2019 <- mmpi_monthly %>% 
  filter(month %in% seq.Date(from = startdate_2019, to = enddate_2019, by = "month")) %>% 
  group_by(pid) %>% 
  filter(any(cohort_2019 == 1))

length(unique(chres_cohort_2019$pid))

chres_cohort_2020 <- mmpi_monthly %>%
  filter(month %in% seq.Date(from = startdate_2020, to = enddate_2020, by = "month")) %>% 
  group_by(pid) %>% 
  filter(any(cohort_2020 == 1))

# Calculate follow up time
chres_cohort_2019 <- chres_cohort_2019 %>% 
  group_by(pid) %>% 
  mutate(max_extractend = max(extractend),
         followup_end = min(deathdate, max_extractend, enddate_2019, na.rm = TRUE)) 

chres_cohort_2019 <- chres_cohort_2019 %>% 
  mutate(days_followup = as.numeric(followup_end - extractstart, 'days'))

chres_cohort_2020 <- chres_cohort_2020 %>% 
  group_by(pid) %>% 
  mutate(max_extractend = max(extractend),
         followup_end = min(deathdate, max_extractend, enddate_2020, na.rm = TRUE)) 

chres_cohort_2020 <- chres_cohort_2020 %>% 
  mutate(days_followup = as.numeric(followup_end - extractstart, 'days'))

chres_cohorts_ts <- chres_cohort_2019 %>% 
  bind_rows(chres_cohort_2020)

# Note: need to be careful to catch deaths, as I've removed 
# extract months after the date of death
# and date of death seems to have been set to the first day of the following extract
chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(censoring_type = case_when(!is.na(deathdate) & deathdate == max_extractend + 1  ~ "death",
                                    followup_end == enddate_2020 | followup_end == enddate_2019 ~ "study end",
                                    followup_end == max_extractend ~ "ch flag end"))

chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(deathmonth_name = ifelse(!is.na(deathdate) & deathdate == max_extractend + 1 &
                                (cohort_2019 == 1 & deathdate >= startdate_2019 & deathdate <= enddate_2019) |
                                   (cohort_2020 == 1 & deathdate >= startdate_2020 & deathdate <= enddate_2020),
                                 as.character(format(deathmonth_filled, "%B")), NA))

# Define categorical variable for care home size 
chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(ch_size = cut(ch_beds, breaks = c(0, 10, 25, 50, 75, Inf), 
                       labels = c("Under 10", "10-24", "25-49", "50-74", "75+"),
                       right = FALSE))

chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(ch_size_res = cut(ch_beds, breaks = c(0, 20, 40, Inf), 
                       labels = c("Under 20", "21-39", "40+"),
                       right = FALSE),
         ch_size_nurs = cut(ch_beds, breaks = c(0, 45, 70, Inf), 
                           labels = c("Under 45", "45-70", "70+"),
                           right = FALSE),
         age_group = cut(age, breaks = c(0, 50, 60, 70, 80, Inf), 
                        labels = c("Under 50", "50-59", "60-69", "70-79", "80+"),
                        right = FALSE) )


# Add IMD information -----------------------------------------------------

chres_cohorts_ts <- chres_cohorts_ts %>% 
  left_join(imd_19_lookup, by = c("lsoa11pidpcd" = "LSOA11"))
# the ones that do not have a match all have missing LSOA

imd_labels <- c('1 least deprived', '2', '3', '4', '5 most deprived')

chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(IMD19_QUINTILE = cut(IMD19_DECILE, breaks = seq(0, 10, by = 2), labels = imd_labels))

chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(IMD19_QUINTILE = fct_explicit_na(IMD19_QUINTILE, na_level = "Missing"))

# Add region information --------------------------------------------------

chres_cohorts_ts <- chres_cohorts_ts %>% 
  left_join(lsoa11_to_region[,c("lsoa11cd", "RGN19CD", "RGN19NM")], 
            by = c("lsoa11pidpcd" = "lsoa11cd"))

chres_cohorts_ts <- chres_cohorts_ts %>% 
  mutate(RGN19NM = fct_explicit_na(RGN19NM, na_level = "Missing"))

chres_cohorts_ts %>%  tabyl(RGN19NM)

# Save longitudinal records -----------------------------------------------

saveRDS(chres_cohorts_ts, str_c(Rds_data_path, 'sprint_4/chres_cohorts_longitudinal.Rds'))


# Add long-term conditions ------------------------------------------------
# these were only counted looking back from the index months Jan 2019 and Jan 2020

chres_cohorts_ts_jan <- chres_cohorts_ts %>% 
  filter(cohort_2019 == 1 | cohort_2020 == 1) 

chres_cohorts_ts_jan <- chres_cohorts_ts_jan %>% 
  left_join(ltcs, by = c("pid", "extractstart"))
 
saveRDS(chres_cohorts_ts_jan, str_c(Rds_data_path, 'sprint_4/chres_cohorts_Jan.Rds'))
  

# Table 1 -----------------------------------------------------------------
# Cohort characteristcs in Jan 2019 and Jan 2020


table1_data <- chres_cohorts_ts_jan %>% 
  mutate(cohort_name = case_when(cohort_2019 == 1 ~ "January 2019",
                            cohort_2020 == 1 ~ "January 2020"))

table1_data <-  table1_data %>% 
  mutate(deathmonth_name = factor(deathmonth_name, levels = c("January", "February", 
                                                      "March", "April",
                                                      "May", "June")))

###########
# Cohort characteristcs in Jan 2019 and Jan 2020
# Split by care home type
# --> in paper
###########

vars_tosummarise_chnursing <- c("sex", "age", "days_followup", "censoring_type", "ch_beds", 
                             "IMD19_QUINTILE", "RGN19NM", "i_charlson_h36",
                             "f_dementia_h36")

cat_vars_tosummarise_chnursing <- c('sex',"censoring_type",
                                 "IMD19_QUINTILE","RGN19NM",
                                 "f_dementia_h36")


table1_residents_chnursing <- CreateTableOne(vars = vars_tosummarise_chnursing, 
                                   data = table1_data, 
                                   strata = c('cohort_name', "ch_nursing"), 
                                   factorVars = cat_vars_tosummarise_chnursing,
                                   test = FALSE, includeNA = TRUE)

table1_residents_chnursing_csv <- print(table1_residents_chnursing, 
                              noSpaces = TRUE,
                              showAllLevels = FALSE) 

write.csv(table1_residents_chnursing_csv, str_c(results_path_sprint4, 'res_chars/Chr_cohorts_chnursing_characteristics.csv'))


# standardised mean difference
table1_residents_chnursing_smd <- ExtractSmd(table1_residents_chnursing) %>%  
  as.data.frame() %>% 
  rownames_to_column(var = "variable") %>% 
  select(variable,
         `Nursing: cohort 2019 vs cohort 2020` = `3 vs 4`,
         `Residential: cohort 2019 vs cohort 2020` = `1 vs 2`) %>% 
  pivot_longer(-variable, names_to = "comparison", values_to = "SMD") %>% 
  mutate(comparison = gsub("_", " ", comparison),
         variable = case_when(variable == "sex" ~ "Female",
                              variable == "RGN19NM" ~ "Region",
                              variable == "IMD19_QUINTILE" ~ "IMD quintile",
                              variable == "i_charlson_h36" ~ "Charlson Index",
                              variable == "f_dementia_h36" ~ "Dementia",
                              variable == "age" ~ "Age",
                              variable == "ch_beds" ~ "Bed capacity",
                              variable == "days_followup" ~ "Follow-up (days)",
                              variable == "censoring_type" ~ "Reason for\nfollow-up ending"),
         variable = factor(variable, levels = c("Female", "Age", "Follow-up (days)", "Reason for\nfollow-up ending",
                                                "Charlson Index", "Dementia",
                                                "IMD quintile", "Region", "Bed capacity")))

write.csv(table1_residents_chnursing_smd, str_c(results_path_sprint4, 'res_chars/Chr_cohorts_chnursing_characteristics_SMD.csv'))


(table1_residents_chnursing_smd %>% 
    ggplot(aes(x = fct_rev(variable), y = SMD*100, group = comparison, color = comparison)) +
    geom_point() +
    coord_flip() +
    ylab("Absolute standardised mean difference (%)") +
    labs(title = "Cohorts of care home residents",
         subtitle = "January") +
    theme_bw() +
    theme(axis.title.y = element_blank(),
          panel.grid.major.x = element_blank(),
          panel.grid.minor.x = element_blank(),
          legend.title = element_blank(),
          legend.position =  "top",
          legend.justification = c(1,0)) +
    scale_color_manual(values = brewer.pal(name ="RdBu", n = 4)[c(2,1,3,4)],
                       guide = guide_legend(ncol = 1)) +
    geom_hline(yintercept = 10, linetype = "dashed", alpha = 0.3)) %>% 
  ggsave(str_c(results_path_sprint4, 'res_chars/Chr_cohorts_chnursing_characteristics_SMD.csv.png'), ., device = "png", 
         dpi = 600, width = 4, height = 4)

# Prepare ID lists for LTC flagging ---------------------------------------

table1_data %>%  
  filter(cohort_2019 == 1) %>% 
  select(pid, extractstart) %>%  
  write_csv(str_c(hist_data_path, "care home paper/cohorts/Cohort_2019.csv"))

table1_data %>%  
  filter(cohort_2020 == 1) %>% 
  select(pid, extractstart) %>%  
  write_csv(str_c(hist_data_path, "care home paper/cohorts/Cohort_2020.csv"))


# Explore care home sizes -------------------------------------------------

table1_data %>% 
  ggplot(aes(x = ch_beds_filled)) +
  geom_histogram() +
  facet_wrap("ch_nursing")
